[IMP] Improves dns_vapor26 routine

+++ dnspro/dns_vapor26.pro
@@ -1,6 +1,5 @@
 
-PRO dns_vapor26, var_3Dlist, isnaps, vdffile
-
+PRO dns_vapor26, var_3Dlist, isnaps, vdffile, do_vdf=do_vdf, do_data=do_data
 
     swap=0
     units="solar"
@@ -27,7 +26,7 @@ PRO dns_vapor26, var_3Dlist, isnaps, vdffile
     ENDIF ELSE BEGIN
        zu=z
     ENDELSE
-    yy  = -reverse(y)
+    yy  = xx
     zz  = -reverse(zu)
     nl  = 2
     ns  = n_elements(isnaps)
@@ -35,44 +34,47 @@ PRO dns_vapor26, var_3Dlist, isnaps, vdffile
     dim = [nx,ny,nz]
     ext = [xx[0],yy[0],zz[0],xx[nx-1],yy[ny-1],zz[nz-1]]
 
-    commandvdf="vdfcreate -level "+strtrim(string(nl),2)+ $
-               " -vars3d "+strjoin(var_3Dlist,':')+ var_2Dlist + $
-               " -dimension "+strjoin(strtrim(string(dim,format="(I4)"),2),'x')+ $
-               " -extents "+strjoin(strtrim(string(ext),2),':')+ $
-               " -numts "+string(nt)+" -periodic 1:1:0 " $
-               + filetimesswitch +   ' '+vdffile
-    SPAWN, commandvdf  
+    IF (KEYWORD_SET(do_vdf)) THEN BEGIN
+       commandvdf="vdfcreate -level "+strtrim(string(nl),2)+ $
+                  " -vars3d "+strjoin(var_3Dlist,':')+ var_2Dlist + $
+                  " -dimension "+strjoin(strtrim(string(dim,format="(I4)"),2),'x')+ $
+                  " -extents "+strjoin(strtrim(string(ext),2),':')+ $
+                  " -numts "+string(nt)+" -periodic 1:1:0 " $
+                  + filetimesswitch +   ' '+vdffile
+       SPAWN, commandvdf
+    ENDIF ELSE print, "No vdf created"
 
-
-    FOR jj=0,ns-1 DO BEGIN
+    IF  (KEYWORD_SET(do_data)) THEN BEGIN
+       FOR jj=0,ns-1 DO BEGIN
        
-       FOR ll=0,N_ELEMENTS(var_3Dlist)-1 DO BEGIN
-          name=var_3Dlist[ll]
-          print, "Loading...  ", name, string(isnaps[jj])
-          dnsvar_name="dnsvar_"+name
-          file_exists=STRLEN(file_which(dnsvar_name+".pro"))
-          IF (file_exists GT 0) THEN BEGIN 
-             CALL_PROCEDURE, dnsvar_name, d, name, isnaps[jj], swap, var, units,$
-                             var_log=var_log
-             IF (var_log) THEN var=alog10(var)
-             IF (abs(min(dz1d)-dz) GT 1e-5) THEN BEGIN
-                FOR i=0,nx-1 DO BEGIN
-                   FOR j=0,ny-1 DO var(i,j,*)=INTERPOL(var(i,j,*),z,zu)
-                ENDFOR
+          FOR ll=0,N_ELEMENTS(var_3Dlist)-1 DO BEGIN
+             name=var_3Dlist[ll]
+             print, "Loading...  ", name, string(isnaps[jj])
+             dnsvar_name="dnsvar_"+name
+             file_exists=STRLEN(file_which(dnsvar_name+".pro"))
+             IF (file_exists GT 0) THEN BEGIN 
+                CALL_PROCEDURE, dnsvar_name, d, name, isnaps[jj], swap, var, units,$
+                                var_log=var_log
+                IF (var_log) THEN var=alog10(var)
+                IF (abs(min(dz1d)-dz) GT 1e-5) THEN BEGIN
+                   index  = findgen(nx)
+                   indey  = findgen(ny)
+                   indez  = interpol(findgen(nz), z, zu)
+                   var = INTERPOLATE(var, index, indey, indez, /grid)
+                ENDIF
+                var=reverse(var,3)
+                var=reverse(var,2)
+                help, var
+                openw,lu,'var.dat',/get_lun
+                spawn, 'ls var.dat'
+                writeu,lu,float(var)
+                close,lu
+                free_lun,lu
+                spawn,'raw2vdf -ts '+STRTRIM(string(isnaps[jj]),2)+' -varname ' +name +' '+vdffile+' var.dat'
+                spawn,'rm -f var.dat'
              ENDIF
-             var=reverse(var,3)
-             var=reverse(var,2)
-             help, var
-             openw,lu,'var.dat',/get_lun
-             spawn, 'ls var.dat'
-             writeu,lu,float(var)
-             close,lu
-             free_lun,lu
-             pmm, var
-             spawn,'raw2vdf -ts '+STRTRIM(string(isnaps[jj]),2)+' -varname ' +name +' '+vdffile+' var.dat'
-             spawn,'rm -f var.dat'
-          ENDIF
+          ENDFOR
        ENDFOR
-    ENDFOR
+    ENDIF ELSE print, "No data created"
 
 END
